+++ contracts/deploy/Create3Deployer.sol
@@ -1,0 +1,37 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity ^0.8.0;
+
+import { Create3 } from './Create3.sol';
+
+contract Create3Deployer {
+    event Deployed(bytes32 indexed bytecodeHash, bytes32 indexed salt, address indexed deployedAddress);
+
+    /**
+     * @dev Deploys a contract using `CREATE3`. The address where the contract
+     * will be deployed can be known in advance via {deployedAddress}.
+     *
+     * The bytecode for a contract can be obtained from Solidity with
+     * `type(contractName).creationCode`.
+     *
+     * Requirements:
+     *
+     * - `bytecode` must not be empty.
+     * - `salt` must have not been used for `bytecode` already by the same `msg.sender`.
+     */
+    function deploy(bytes calldata bytecode, bytes32 salt) external returns (address deployedAddress_) {
+        bytes32 deploySalt = keccak256(abi.encode(msg.sender, salt));
+        deployedAddress_ = Create3.deploy(deploySalt, bytecode);
+
+        emit Deployed(keccak256(bytecode), salt, deployedAddress_);
+    }
+
+    /**
+     * @dev Returns the address where a contract will be stored if deployed via {deploy} or {deployAndInit} by `sender`.
+     * Any change in the `bytecode`, `sender`, or `salt` will result in a new destination address.
+     */
+    function deployedAddress(address sender, bytes32 salt) external view returns (address) {
+        bytes32 deploySalt = keccak256(abi.encode(sender, salt));
+        return Create3.deployedAddress(deploySalt, address(this));
+    }
+}
