+++ contracts/interfaces/IOwnable.sol
@@ -1,0 +1,22 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity ^0.8.0;
+
+// General interface for upgradable contracts
+interface IOwnable {
+    error NotOwner();
+    error InvalidOwner();
+
+    event OwnershipTransferStarted(address indexed newOwner);
+    event OwnershipTransferred(address indexed newOwner);
+
+    // Get current owner
+    function owner() external view returns (address);
+
+    // Get pending ownership transfer
+    function pendingOwner() external view returns (address);
+
+    function transferOwnership(address newOwner) external;
+
+    function acceptOwnership() external;
+}
