+++ contracts/interfaces/IProxy.sol
@@ -1,0 +1,15 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity ^0.8.0;
+
+// General interface for upgradable contracts
+interface IProxy {
+    error InvalidOwner();
+    error InvalidImplementation();
+    error SetupFailed();
+    error AlreadyInitialized();
+
+    function implementation() external view returns (address implementation_);
+
+    function setup(bytes calldata setupParams) external;
+}
