+++ contracts/test/ProxyTest.sol
@@ -5,6 +5,12 @@ pragma solidity 0.8.9;
 import { Proxy } from '../upgradable/Proxy.sol';
 
 contract ProxyTest is Proxy {
+    constructor(
+        address implementationAddress,
+        address owner,
+        bytes memory setupParams
+    ) Proxy(implementationAddress, owner, setupParams) {}
+
     function contractId() internal pure override returns (bytes32) {
         return keccak256('test');
     }
