+++ contracts/utils/Ownable.sol
@@ -1,0 +1,53 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity ^0.8.0;
+
+import '../interfaces/IUpgradable.sol';
+
+abstract contract Ownable is IOwnable {
+    // keccak256('owner')
+    bytes32 internal constant _OWNER_SLOT = 0x02016836a56b71f0d02689e69e326f4f4c1b9057164ef592671cf0d37c8040c0;
+    // keccak256('ownership-transfer')
+    bytes32 internal constant _OWNERSHIP_TRANSFER_SLOT =
+        0x9855384122b55936fbfb8ca5120e63c6537a1ac40caf6ae33502b3c5da8c87d1;
+
+    modifier onlyOwner() {
+        if (owner() != msg.sender) revert NotOwner();
+
+        _;
+    }
+
+    function owner() public view returns (address owner_) {
+        // solhint-disable-next-line no-inline-assembly
+        assembly {
+            owner_ := sload(_OWNER_SLOT)
+        }
+    }
+
+    function pendingOwner() public view returns (address owner_) {
+        // solhint-disable-next-line no-inline-assembly
+        assembly {
+            owner_ := sload(_OWNERSHIP_TRANSFER_SLOT)
+        }
+    }
+
+    function transferOwnership(address newOwner) external virtual onlyOwner {
+        emit OwnershipTransferStarted(newOwner);
+        // solhint-disable-next-line no-inline-assembly
+        assembly {
+            sstore(_OWNERSHIP_TRANSFER_SLOT, newOwner)
+        }
+    }
+
+    function acceptOwnership() external virtual {
+        address newOwner = pendingOwner();
+        if (newOwner != msg.sender) revert InvalidOwner();
+
+        emit OwnershipTransferred(newOwner);
+        // solhint-disable-next-line no-inline-assembly
+        assembly {
+            sstore(_OWNERSHIP_TRANSFER_SLOT, 0)
+            sstore(_OWNER_SLOT, newOwner)
+        }
+    }
+}
