+++ package.json
@@ -1,12 +1,12 @@
 {
   "name": "@axelar-network/axelar-gmp-sdk-solidity",
-  "version": "3.0.0",
+  "version": "3.1.0",
   "description": "Solidity GMP SDK and utilities provided by Axelar for cross-chain development",
   "main": "index.js",
   "scripts": {
     "test": "npx hardhat test",
     "build": "rm -rf artifacts && hardhat compile && npm run copy:artifacts",
-    "copy:artifacts": "rm -rf dist && mkdir dist && cp artifacts/contracts/ConstAddressDeployer.sol/ConstAddressDeployer.json artifacts/contracts/interfaces/IUpgradable.sol/IUpgradable.json dist/",
+    "copy:artifacts": "rm -rf dist && mkdir dist && cp artifacts/contracts/express/ExpressProxyDeployer.sol/ExpressProxyDeployer.json artifacts/contracts/deploy/Create3Deployer.sol/Create3Deployer.json artifacts/contracts/interfaces/IUpgradable.sol/IUpgradable.json dist/",
     "lint": "solhint 'contracts/**/*.sol' && eslint 'scripts/**/*.js' 'test/**/*.js'",
     "prettier": "prettier --write 'contracts/**/*.sol' 'test/**/*.js' 'scripts/**/*.js' '*.js'",
     "flatten": "sh scripts/flatten-contracts.sh"
