+++ test/NftLinker.js
@@ -13,7 +13,7 @@ const { expect } = chai;
 const AxelarGateway = require('../artifacts/contracts/test/MockGateway.sol/MockGateway.json');
 const GasService = require('../artifacts/contracts/test/MockGasService.sol/MockGasService.json');
 const ERC721MintableBurnable = require('../artifacts/contracts/test/ERC721MintableBurnable.sol/ERC721MintableBurnable.json');
-const ConstAddressDeployer = require('../dist/ConstAddressDeployer.json');
+const Create3Deployer = require('../dist/Create3Deployer.json');
 const NftLinkerProxy = require('../artifacts/contracts/nft-linker/NftLinkerProxy.sol/NftLinkerProxy.json');
 const NftLinkerLockUnlock = require('../artifacts/contracts/nft-linker/NftLinkerLockUnlock.sol/NftLinkerLockUnlock.json');
 const NftLinkerMintBurn = require('../artifacts/contracts/nft-linker/NftLinkerMintBurn.sol/NftLinkerMintBurn.json');
@@ -53,10 +53,7 @@ describe('NftLinker', () => {
   beforeEach(async () => {
     gateway = await deployContract(ownerWallet, AxelarGateway);
     gasService = await deployContract(ownerWallet, GasService);
-    constAddressDeployer = await deployContract(
-      ownerWallet,
-      ConstAddressDeployer,
-    );
+    constAddressDeployer = await deployContract(ownerWallet, Create3Deployer);
 
     token = await deployContract(ownerWallet, ERC721MintableBurnable, [
       tokenName,
