+++ test/TokenLinker.js
@@ -13,7 +13,7 @@ const { expect } = chai;
 const AxelarGateway = require('../artifacts/contracts/test/MockGateway.sol/MockGateway.json');
 const GasService = require('../artifacts/contracts/test/MockGasService.sol/MockGasService.json');
 const ERC20MintableBurnable = require('../artifacts/contracts/test/ERC20MintableBurnable.sol/ERC20MintableBurnable.json');
-const ConstAddressDeployer = require('../dist/ConstAddressDeployer.json');
+const Create3Deployer = require('../dist/Create3Deployer.json');
 const TokenLinkerProxy = require('../artifacts/contracts/token-linker/TokenLinkerProxy.sol/TokenLinkerProxy.json');
 const TokenLinkerLockUnlock = require('../artifacts/contracts/token-linker/TokenLinkerLockUnlock.sol/TokenLinkerLockUnlock.json');
 const TokenLinkerMintBurn = require('../artifacts/contracts/token-linker/TokenLinkerMintBurn.sol/TokenLinkerMintBurn.json');
@@ -55,10 +55,7 @@ describe('TokenLinker', () => {
   beforeEach(async () => {
     gateway = await deployContract(ownerWallet, AxelarGateway);
     gasService = await deployContract(ownerWallet, GasService);
-    constAddressDeployer = await deployContract(
-      ownerWallet,
-      ConstAddressDeployer,
-    );
+    constAddressDeployer = await deployContract(ownerWallet, Create3Deployer);
 
     token = await deployContract(ownerWallet, ERC20MintableBurnable, [
       tokenName,
